+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.0.0] - 2021-06-14
 ### Added
 - Rearrange assumption strategy
 - `TrackingAssumption.from_cell` method
@@ -13,15 +14,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `FusionParallelSpecFinder`, `FusionIsomorphism` and `FusionBijection` subclasses
   to handle bijection involving fusion and assumption strategies.
 - a `insertion_point_row_and_col_placements` pack in `TileScopePack`
+- `TrackedSearcher` and `ForgetTrackedSearcher` that prioritise expanding classes
+  with assumptions when the underlying tiling was expanded
 
 ### Changed
-- Updated to use comb_spec_searcher 3.0.0
+- Updated to use comb_spec_searcher 4.0.0
 - Using `strategy_pack.make_fusion` or `strategy_pack.make_tracked` now automatically
   adds `RearrangeAssumptionFactory` with `apply_first=True` if `tracked=True`
 - Moved several local function in `RearrangeConstructor` outward to make it pickleable
 - Fixed isolated fusion bug
 - Fusion is now a one-way strategy.
 - Added length to name for `all_the_strategies` pack
+- pack for locally factorable now account for the basis.
 
 ### Fixed
 - Removed a redundant `Tiling.is_empty` check in the fusion algorithm.